% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{analyze_own_data}
\alias{analyze_own_data}
\title{Analyze User Data}
\usage{
analyze_own_data(data, plot = TRUE)
}
\arguments{
\item{data}{Data frame containing columns: Drug_Name, Cell_Line, Dose, Viability.}

\item{plot}{Logical, whether to generate plots.}
}
\value{
A list containing metrics and plots.
}
\description{
Takes a dataframe of user data, fits curves, calculates metrics, and optionally plots the result.
}
