% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{calculate_all}
\alias{calculate_all}
\title{Calculate All Metrics}
\usage{
calculate_all(
  dose,
  inhibition_percent,
  dataset,
  drug_name,
  DSS_type = 2,
  min_dose_range = NULL,
  max_dose_range = NULL
)
}
\arguments{
\item{dose}{Numeric vector of doses}

\item{inhibition_percent}{Numeric vector of inhibition}

\item{dataset}{Dataset name}

\item{drug_name}{Drug name}

\item{DSS_type}{DSS type (default 2)}

\item{min_dose_range}{Optional min dose}

\item{max_dose_range}{Optional max dose}
}
\description{
Calculate All Metrics
}
