% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{calculate_dss}
\alias{calculate_dss}
\title{Drug Sensitivity Score (DSS) Calculation}
\usage{
calculate_dss(
  ic50,
  slope,
  max,
  min.conc.tested,
  max.conc.tested,
  y = 10,
  DSS.type = 2,
  concn_scale = 1e-09
)
}
\description{
Drug Sensitivity Score (DSS) Calculation
}
