% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{calculate_msf_variants}
\alias{calculate_msf_variants}
\title{Calculate Multi-Source Feature (MSF) Variants}
\usage{
calculate_msf_variants(
  ic50,
  slope,
  max_response,
  min_conc,
  max_conc,
  residuals = NULL,
  observed_values = NULL,
  ic50_abs = NULL,
  nls_model = NULL,
  min_activity = 10,
  min_dose_range = NULL,
  max_dose_range = NULL
)
}
\description{
Calculate Multi-Source Feature (MSF) Variants
}
