% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{get_metrics}
\alias{get_metrics}
\title{Get Metrics for a Drug-Cell Line Pair}
\usage{
get_metrics(drug, cell_line, dose, inhibition)
}
\description{
Wraps the analysis function for a single pair input.
}
